backend/requirements.txt

fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
azure-storage-blob==12.17.0
azure-identity==1.13.0
azure-storage-file-datalake==12.11.0
pydantic==1.10.8
python-multipart==0.0.6
requests==2.31.0
pandas==2.0.2
pyarrow==12.0.1
deltalake==0.10.0
cors==1.0.1
